Struct secret_integers::I16

pub struct I16(pub i16);

Implementations

impl I16

pub fn classify<T: Into<i16>>(x: T) -> Self

pub fn declassify(self) -> i16

Warning: use with caution, breaks the constant-time guarantee.

pub fn zero() -> Self

pub fn one() -> Self

pub fn ones() -> Self

pub fn from_le_bytes(bytes: &[U8]) -> Vec<I16>

pub fn to_le_bytes(ints: &[I16]) -> Vec<U8>

pub fn from_be_bytes(bytes: &[U8]) -> Vec<I16>

pub fn to_be_bytes(ints: &[I16]) -> Vec<U8>

pub fn max_value() -> I16

impl I16

pub fn checked_add(self, rhs: Self) -> Self

Warning: panics when overflow.

impl I16

pub fn checked_sub(self, rhs: Self) -> Self

Warning: panics when overflow.

impl I16

pub fn checked_mul(self, rhs: Self) -> Self

Warning: panics when overflow.

impl I16

pub fn rotate_left(self, rotval: u32) -> Self

pub fn rotate_right(self, rotval: u32) -> Self

impl I16

Constant-time comparison operators

pub fn comp_eq(self, rhs: Self) -> Self

pub fn comp_ne(self, rhs: Self) -> Self

Produces a new integer which is all ones if the first argument is different from the second argument, and all zeroes otherwise.

pub fn comp_gte(self, rhs: Self) -> Self

Produces a new integer which is all ones if the first argument is greater than or equal to the second argument, and all zeroes otherwise. With inspiration from

pub fn comp_gt(self, rhs: Self) -> Self

Produces a new integer which is all ones if the first argument is strictly greater than the second argument, and all zeroes otherwise.

pub fn comp_lte(self, rhs: Self) -> Self

Produces a new integer which is all ones if the first argument is less than or equal to the second argument, and all zeroes otherwise.

pub fn comp_lt(self, rhs: Self) -> Self

Produces a new integer which is all ones if the first argument is strictly less than the second argument, and all zeroes otherwise.